# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clerkship API - ClinicAI UNAB
  description: '**API REST de la plataforma clínica Clerkship / ClinicAI UNAB** para
    el entrenamiento y evaluación

    del proceso de razonamiento diagnóstico en estudiantes de medicina y ciencias
    de la salud.


    El backend combina:

    * **PostgreSQL** para la persistencia relacional transaccional (usuarios, cursos,
    artículos, carpetas y evaluaciones).

    * **MongoDB** para almacenamiento documental semiestructurado (transcripción de
    chats clínicos, notas de voz y documentos en base64).

    * **JWT (JSON Web Tokens)** con control de acceso basado en roles (`STUDENT`,
    `TEACHER`, `ADMIN`).

    '
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo Clerkship
    url: https://github.com/Steven08Ar/Clerkship
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
- url: http://localhost:5000
  description: Servidor local de desarrollo Flask
...
paths:
...
  /api/usuarios/{id}:
    get:
      tags:
      - Usuarios
      summary: Perfil público de un usuario
      description: Retorna la información pública de un usuario por su UUID.
      operationId: getUserById
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Información pública del usuario obtenida con éxito
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/comunidad/posts:
    get:
      tags:
      - Comunidad
      summary: Listar publicaciones de la comunidad
      operationId: listCommunityPosts
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommunityPostResponse'
          description: Publicaciones de la comunidad obtenidas con éxito
    post:
      tags:
      - Comunidad
      summary: Crear nueva publicación
      operationId: createCommunityPost
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostResponse'
          description: Publicación creada exitosamente
  /api/comunidad/posts/{id}:
    delete:
      tags:
      - Comunidad
      summary: Eliminar publicación propia
      operationId: deleteCommunityPost
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Publicación eliminada
...
  /api/admin/estadisticas:
    get:
      tags:
      - Administración
      summary: Estadísticas generales de la plataforma
      operationId: adminEstadisticas
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Correcto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/usuarios:
    get:
      tags:
      - Administración
      summary: Lista y busca usuarios
      operationId: adminListarUsuarios
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Correcto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/usuarios/{user_id}:
    patch:
      tags:
      - Administración
      summary: Cambia el rol o activa/desactiva una cuenta
      operationId: adminActualizarUsuario
      security:
      - BearerAuth: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema: {type: string}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
                activo: {type: boolean}
      responses:
        '200':
          description: Correcto
        '400':
          description: Datos inválidos
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/comunidad/posts:
    get:
      tags:
      - Administración
      summary: Lista publicaciones de la comunidad
      operationId: adminListarPosts
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Correcto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/comunidad/posts/{post_id}:
    delete:
      tags:
      - Administración
      summary: Borra cualquier publicación
      operationId: adminBorrarPost
      security:
      - BearerAuth: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Correcto
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/comunidad/comentarios/{comentario_id}:
    delete:
      tags:
      - Administración
      summary: Borra cualquier comentario
      operationId: adminBorrarComentario
      security:
      - BearerAuth: []
      parameters:
      - name: comentario_id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Correcto
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/biblioteca:
    get:
      tags:
      - Administración
      summary: Lista los recursos de la biblioteca
      operationId: adminListarBiblioteca
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Correcto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/biblioteca/{article_id}:
    delete:
      tags:
      - Administración
      summary: Borra un recurso de la biblioteca
      operationId: adminBorrarArticulo
      security:
      - BearerAuth: []
      parameters:
      - name: article_id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Correcto
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/feedback/{pestana}:
    get:
      tags:
      - Administración
      summary: Lista las respuestas de retroalimentación de una pestaña
      operationId: adminListarFeedback
      security:
      - BearerAuth: []
      parameters:
      - name: pestana
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Correcto
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/admin/auditoria:
    get:
      tags:
      - Admin
      summary: Registro de auditoría (solo administradores), del más reciente al más viejo
      description: >-
        Acciones registradas: GRADE_CHANGE, SUBMISSION_DELETE, SUBMISSION_RESTORE, STUDENT_ENROLL,
        STUDENT_REMOVE, ENROLLMENT_REQUEST_REJECT, COURSE_DELETE, COURSE_RESTORE, COURSE_PURGE,
        ENROLLMENT_CODE_CHANGE, ENROLLMENT_MODE_CHANGE, GRADE_POLICY_CHANGE, USER_ROLE_CHANGE,
        USER_ACTIVE_CHANGE. El registro solo se puede leer: no hay rutas para editarlo ni borrarlo.
      operationId: listarAuditoria
      security:
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - {name: action, in: query, required: false, schema: {type: string}}
      - {name: entity_type, in: query, required: false, schema: {type: string}}
      - {name: entity_id, in: query, required: false, schema: {type: string}}
      - {name: user_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Correcto
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Paginacion'
                - type: object
                  properties:
                    registros:
                      type: array
                      items:
                        type: object
                        properties:
                          id: {type: string, format: uuid}
                          user_id: {type: string, format: uuid, nullable: true}
                          user_name: {type: string, nullable: true}
                          action: {type: string}
                          entity_type: {type: string}
                          entity_id: {type: string, nullable: true}
                          old_value: {type: object, nullable: true}
                          new_value: {type: object, nullable: true}
                          ip: {type: string, nullable: true}
                          created_at: {type: string, format: date-time}
        '400':
          description: user_id inválido
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/admin/validacion/{pestana}:
    get:
      tags:
      - Administración
      summary: Lista las respuestas de validación por expertos de una pestaña
      operationId: adminListarValidacion
      security:
      - BearerAuth: []
      parameters:
      - name: pestana
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Correcto
        '404':
          description: No encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/admin/tokens:
    get:
      tags:
      - Administración
      summary: Uso de tokens de los 3 agentes del simulador, con resumen general
      operationId: adminEstadisticasTokens
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Correcto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/feedback/inicio:
    post:
      tags:
      - Retroalimentación
      summary: Formulario de retroalimentación de la pestaña Inicio
      description: Guarda las calificaciones (1 a 5) en la tabla feedback_inicio. El usuario sale del JWT. Rol requerido cualquier usuario autenticado.
      operationId: feedbackInicio
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_usuario, claridad_navegacion, informacion_util, accesos_rapidos, velocidad_carga, diseno_visual, legibilidad_textos, orden_menus, facilidad_uso_movil, confianza_plataforma, claridad_mensajes, utilidad_avisos, recomendaria_plataforma]
              properties:
                claridad_navegacion: {type: integer, minimum: 1, maximum: 5}
                informacion_util: {type: integer, minimum: 1, maximum: 5}
                accesos_rapidos: {type: integer, minimum: 1, maximum: 5}
                velocidad_carga: {type: integer, minimum: 1, maximum: 5}
                diseno_visual: {type: integer, minimum: 1, maximum: 5}
                legibilidad_textos: {type: integer, minimum: 1, maximum: 5}
                orden_menus: {type: integer, minimum: 1, maximum: 5}
                facilidad_uso_movil: {type: integer, minimum: 1, maximum: 5}
                confianza_plataforma: {type: integer, minimum: 1, maximum: 5}
                claridad_mensajes: {type: integer, minimum: 1, maximum: 5}
                utilidad_avisos: {type: integer, minimum: 1, maximum: 5}
                recomendaria_plataforma: {type: integer, minimum: 1, maximum: 5}
                comentario: {type: string, maxLength: 1000, nullable: true}
      responses:
        '201':
          description: Respuesta guardada
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/feedback/casos:
    post:
      tags:
      - Retroalimentación
      summary: Formulario de retroalimentación de la pestaña Casos clínicos
      description: Guarda las calificaciones (1 a 5) en la tabla feedback_casos. El usuario sale del JWT. Rol requerido cualquier usuario autenticado.
      operationId: feedbackCasos
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_usuario, realismo_caso, calidad_paciente_virtual, claridad_instrucciones, dificultad_percibida, coherencia_sintomas, coherencia_examenes, utilidad_educativa, retroalimentacion_clara, tiempo_adecuado, variedad_casos, realismo_examenes, satisfaccion_general]
              properties:
                realismo_caso: {type: integer, minimum: 1, maximum: 5}
                calidad_paciente_virtual: {type: integer, minimum: 1, maximum: 5}
                claridad_instrucciones: {type: integer, minimum: 1, maximum: 5}
                dificultad_percibida: {type: integer, minimum: 1, maximum: 5}
                coherencia_sintomas: {type: integer, minimum: 1, maximum: 5}
                coherencia_examenes: {type: integer, minimum: 1, maximum: 5}
                utilidad_educativa: {type: integer, minimum: 1, maximum: 5}
                retroalimentacion_clara: {type: integer, minimum: 1, maximum: 5}
                tiempo_adecuado: {type: integer, minimum: 1, maximum: 5}
                variedad_casos: {type: integer, minimum: 1, maximum: 5}
                realismo_examenes: {type: integer, minimum: 1, maximum: 5}
                satisfaccion_general: {type: integer, minimum: 1, maximum: 5}
                comentario: {type: string, maxLength: 1000, nullable: true}
      responses:
        '201':
          description: Respuesta guardada
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/feedback/historial:
    post:
      tags:
      - Retroalimentación
      summary: Formulario de retroalimentación de la pestaña Historial
      description: Guarda las calificaciones (1 a 5) en la tabla feedback_historial. El usuario sale del JWT. Rol requerido cualquier usuario autenticado.
      operationId: feedbackHistorial
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_usuario, claridad_puntajes, utilidad_recomendacion, facilidad_revision, detalle_evaluacion, comprension_dimensiones, progreso_visible, utilidad_para_estudiar, filtros_utiles, exportacion_util, satisfaccion_general]
              properties:
                claridad_puntajes: {type: integer, minimum: 1, maximum: 5}
                utilidad_recomendacion: {type: integer, minimum: 1, maximum: 5}
                facilidad_revision: {type: integer, minimum: 1, maximum: 5}
                detalle_evaluacion: {type: integer, minimum: 1, maximum: 5}
                comprension_dimensiones: {type: integer, minimum: 1, maximum: 5}
                progreso_visible: {type: integer, minimum: 1, maximum: 5}
                utilidad_para_estudiar: {type: integer, minimum: 1, maximum: 5}
                filtros_utiles: {type: integer, minimum: 1, maximum: 5}
                exportacion_util: {type: integer, minimum: 1, maximum: 5}
                satisfaccion_general: {type: integer, minimum: 1, maximum: 5}
                comentario: {type: string, maxLength: 1000, nullable: true}
      responses:
        '201':
          description: Respuesta guardada
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/feedback/biblioteca:
    post:
      tags:
      - Retroalimentación
      summary: Formulario de retroalimentación de la pestaña Biblioteca
      description: Guarda las calificaciones (1 a 5) en la tabla feedback_biblioteca. El usuario sale del JWT. Rol requerido cualquier usuario autenticado.
      operationId: feedbackBiblioteca
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_usuario, calidad_contenido, facilidad_busqueda, utilidad_recursos, actualidad_contenido, claridad_descripciones, variedad_recursos, facilidad_descarga, organizacion_por_temas, calidad_lectura, utilidad_para_estudio, satisfaccion_general]
              properties:
                calidad_contenido: {type: integer, minimum: 1, maximum: 5}
                facilidad_busqueda: {type: integer, minimum: 1, maximum: 5}
                utilidad_recursos: {type: integer, minimum: 1, maximum: 5}
                actualidad_contenido: {type: integer, minimum: 1, maximum: 5}
                claridad_descripciones: {type: integer, minimum: 1, maximum: 5}
                variedad_recursos: {type: integer, minimum: 1, maximum: 5}
                facilidad_descarga: {type: integer, minimum: 1, maximum: 5}
                organizacion_por_temas: {type: integer, minimum: 1, maximum: 5}
                calidad_lectura: {type: integer, minimum: 1, maximum: 5}
                utilidad_para_estudio: {type: integer, minimum: 1, maximum: 5}
                satisfaccion_general: {type: integer, minimum: 1, maximum: 5}
                comentario: {type: string, maxLength: 1000, nullable: true}
      responses:
        '201':
          description: Respuesta guardada
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token de acceso JWT obtenido al iniciar sesión (`/api/auth/login`)
        o registrarse.

        Debe enviarse en la cabecera: `Authorization: Bearer <token>`.

        '
  responses:
...
    UnauthorizedError:
      description: Token de autenticación ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
            message: Token ausente o expirado
            status_code: 401
...
    NotFoundError:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not Found
            message: El recurso solicitado no fue encontrado.
            status_code: 404
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: sarias202
        email:
          type: string
          format: email
        first_name:
          type: string
          example: Santiago
        last_name:
          type: string
          example: Arias
        role:
          type: string
          enum:
          - STUDENT
          - TEACHER
          - ADMIN
          example: STUDENT
        email_verified:
          type: boolean
          example: true
        avatar_svg:
          type: string
          nullable: true
...
    CreatePostRequest:
      type: object
      required:
      - title
      - content
      properties:
        title:
          type: string
          example: 'Discusión: Marcadores tempranos en Pancreatitis Aguda'
        content:
          type: string
    CommunityPostResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        author_name:
          type: string
        title:
          type: string
        content:
          type: string
        likes_count:
          type: integer
          example: 7
        comments_count:
          type: integer
          example: 3
        created_at:
          type: string
          format: date-time
...